refactoring + small fixes
