+++ client/src/modules/Result/index.svelte
@@ -1,1 +1,6 @@
+<script lang="ts">
+
+	import { state } from '../../stores/states';
+</script>
+
 <div>Test completed!</div>
