+++ client/src/stores/states.ts
@@ -1,3 +1,4 @@
 import { writable } from 'svelte/store';
+import { States } from '../utils/constants';
 
-export const state = writable("typing")
+export const state = writable(States.typing)
