+++ client/src/utils/constants.ts
@@ -1,6 +1,11 @@
-export const CHAR_ERROR = "text-red-400"
-export const CHAR_CORRECT = "text-gray-300"
+export enum Class {
+	CHAR_ERROR="text-red-400",
+	CHAR_CORRECT = "text-gray-300",
+	COLOR_TRANSITION = "transition-colors",
+	WORD_INCORRECT = "underline-error"
+}
 
-export const COLOR_TRANSITION = "transition-colors"
-
-export const WORD_INCORRECT = "underline-error"
+export enum States {
+	typing = "typing",
+	result = "result"
+}
